requirements.txt
requests
PyYAML
click==3.3
setuptools
dciclient
jinja2
